Connect server - send and receive data which return from server
